+++ src/BlockTreeWithRates.h
@@ -22,30 +22,15 @@ public:
   template<typename RngType = std::mt19937_64>
   void handleEvent (Event &ev, CategorySampler& sampler, RngType &rng) {
     if (ev.length == 0) return;
-    // std::cout << _avlTree->print_avl();
-
-    // std::cout << "Handling event: type=" << (ev.type == INSERTION ? "INSERTION" : "DELETION") 
-    // << " position=" << ev.position 
-    // << " length=" << ev.length << std::endl;
-
     if(!_avlTree->handle_event(ev, sampler, rng)) {
       throw std::out_of_range("event_position exceeds sequence");
     }
-    // if (!_avlTree->validate_rate_integrity()) {
-    //       std::cout << "Handling event: type=" << (ev.type == INSERTION ? "INSERTION" : "DELETION") 
-    //           << " position=" << ev.position 
-    //           << " length=" << ev.length << std::endl;
-    //   errorMsg::reportError("Rate integrity is broken\nExiting...");
-    // }
   }
 
   std::string printTree () {
     return _avlTree->print_avl();
   }
 
-  // BlockList getBlockList () {
-  //   return _avlTree->get_blocklist();
-  // }
 
   TreeType::iterator begin () {
     return _avlTree->begin();
@@ -71,7 +56,7 @@ public:
     _avlTree->clear();
   }
 
-  void initTree(int first_block_size, const std::vector<size_t>& rateCategories){
+  void initTree(int first_block_size, const std::vector<uint8_t>& rateCategories){
     _avlTree->clear();
     _avlTree->init_tree(first_block_size + 1, rateCategories);
   }
